fix(react-core): allow Sizer to update size on cDU
